+++ genericsuite_ai/config/config.py
@@ -45,22 +45,6 @@ class Config(ConfigSuperClass):
             # 'AI_TECHNOLOGY', 'openai'
         )
 
-        self.EMBEDDINGS_ENGINE = self.get_env(
-            'EMBEDDINGS_ENGINE', 'openai'
-            # 'EMBEDDINGS_ENGINE', 'clarifai'
-            # 'EMBEDDINGS_ENGINE', 'bedrock'
-            # 'EMBEDDINGS_ENGINE', 'huggingface'
-            # 'EMBEDDINGS_ENGINE', 'cohere'
-            # 'EMBEDDINGS_ENGINE', 'ollama'
-        )
-
-        self.VECTOR_STORE_ENGINE = self.get_env(
-            'VECTOR_STORE_ENGINE', 'FAISS'
-            # 'VECTOR_STORE_ENGINE', 'clarifai'
-            # 'VECTOR_STORE_ENGINE', 'mongo'
-            # 'VECTOR_STORE_ENGINE', 'vectara'
-        )
-
         self.LANGCHAIN_DEFAULT_MODEL = self.get_env(
             'LANGCHAIN_DEFAULT_MODEL', 'chat_openai'
             # 'LANGCHAIN_DEFAULT_MODEL', 'anthropic'
@@ -71,6 +55,8 @@ class Config(ConfigSuperClass):
             # 'LANGCHAIN_DEFAULT_MODEL', 'gemini'
             # 'LANGCHAIN_DEFAULT_MODEL', 'clarifai'
             # 'LANGCHAIN_DEFAULT_MODEL', 'bedrock'
+            # 'LANGCHAIN_DEFAULT_MODEL', 'ollama'
+            # 'LANGCHAIN_DEFAULT_MODEL', 'aimlapi'
         )
 
         self.AI_VISION_TECHNOLOGY = self.get_env(
@@ -97,6 +83,48 @@ class Config(ConfigSuperClass):
         self.AI_TEXT_TO_AUDIO_TECHNOLOGY = self.get_env(
             'AI_TEXT_TO_AUDIO_TECHNOLOGY', 'openai'
             # 'AI_TEXT_TO_AUDIO_TECHNOLOGY', 'clarifai'
+        )
+
+        self.EMBEDDINGS_ENGINE = self.get_env(
+            'EMBEDDINGS_ENGINE', 'openai'
+            # 'EMBEDDINGS_ENGINE', 'clarifai'
+            # 'EMBEDDINGS_ENGINE', 'bedrock'
+            # 'EMBEDDINGS_ENGINE', 'huggingface'
+            # 'EMBEDDINGS_ENGINE', 'cohere'
+            # 'EMBEDDINGS_ENGINE', 'ollama'
+        )
+
+        self.VECTOR_STORE_ENGINE = self.get_env(
+            'VECTOR_STORE_ENGINE', 'FAISS'
+            # 'VECTOR_STORE_ENGINE', 'clarifai'
+            # 'VECTOR_STORE_ENGINE', 'mongo'
+            # 'VECTOR_STORE_ENGINE', 'vectara'
+        )
+
+        self.AI_MODEL_ALLOW_SYSTEM_MSG = self.get_env(
+            'AI_MODEL_ALLOW_SYSTEM_MSG', '')
+
+        self.AI_MODEL_ALLOW_TOOLS = self.get_env(
+            'AI_MODEL_ALLOW_TOOLS', '')
+
+        self.AI_MODEL_NEED_PREAMBLE = self.get_env(
+            'AI_MODEL_NEED_PREAMBLE', '')
+
+        self.AI_PREAMBLE_MODEL_DEFAULT_TYPE = self.get_env(
+            'AI_PREAMBLE_MODEL_DEFAULT_TYPE', 'chat_openai')
+
+        self.AI_PREAMBLE_MODEL_DEFAULT_MODEL = self.get_env(
+            'AI_PREAMBLE_MODEL_DEFAULT_MODEL', 'gpt-4o-mini')
+
+        self.AI_PREAMBLE_MODEL_BASE_CONF = self.get_env(
+            'AI_PREAMBLE_MODEL_BASE_CONF',
+            '{"o1-mini": {"model_type": "chat_openai", "model_name": ' +
+            '"gpt-4o-mini"}, "o1-preview": {"model_type": "chat_openai",' +
+            ' "model_name": "gpt-4o-mini"}}'
+        )
+
+        self.AI_PREAMBLE_MODEL_CUSTOM_CONF = self.get_env(
+            'AI_PREAMBLE_MODEL_CUSTOM_CONF', ''
         )
 
         self.AI_ADDITIONAL_MODELS = self.get_env(
@@ -249,12 +277,12 @@ class Config(ConfigSuperClass):
         self.AWS_BEDROCK_MODEL_ID = self.get_env(
             'AWS_BEDROCK_MODEL_ID',
             "amazon.titan-text-premier-v1:0"
-            "amazon.titan-text-express-v1"
-            "ai21.jamba-instruct-v1:0"
-            "anthropic.claude-3-haiku-20240307-v1:0"
-            "anthropic.claude-3-opus-20240229-v1:0"
-            "anthropic.claude-3-sonnet-20240229-v1:0"
-            "anthropic.claude-3-5-sonnet-20240229-v1:0"
+            # "amazon.titan-text-express-v1"
+            # "ai21.jamba-instruct-v1:0"
+            # "anthropic.claude-3-haiku-20240307-v1:0"
+            # "anthropic.claude-3-opus-20240229-v1:0"
+            # "anthropic.claude-3-sonnet-20240229-v1:0"
+            # "anthropic.claude-3-5-sonnet-20240229-v1:0"
         )
 
         self.AWS_BEDROCK_IMAGE_GEN_MODEL_ID = self.get_env(
@@ -436,12 +464,24 @@ class Config(ConfigSuperClass):
         # AI/ML API
 
         self.AIMLAPI_API_KEY = self.get_env('AIMLAPI_API_KEY', "")
+
+        # Reference:
+        # https://docs.aimlapi.com/api-overview/model-database/text-models
         self.AIMLAPI_MODEL_NAME = self.get_env(
             'AIMLAPI_MODEL_NAME',
+            # https://aimlapi.com/models/openai-o1-preview-api
             "o1-mini"
             # "o1-preview"
         )
-        self.AIMLAPI_TEMPERATURE = self.get_env('AIMLAPI_TEMPERATURE', '1')
+
+        self.AIMLAPI_TEMPERATURE = self.get_env(
+            'AIMLAPI_TEMPERATURE', '1')
+
+        self.AIMLAPI_MAX_TOKENS = self.get_env(
+            'AIMLAPI_MAX_TOKENS',
+            # "o1-preview" model supports at most 32768 completion tokens
+            '32768'
+        )
 
         self.AIMLAPI_BASE_URL = self.get_env(
             'AIMLAPI_BASE_URL',
@@ -478,9 +518,22 @@ class Config(ConfigSuperClass):
 
         # Ollama credentials and other parameters
 
-        self.OLLAMA_MODEL = self.get_env('OLLAMA_MODEL', "llama:7b")
+        self.OLLAMA_MODEL = self.get_env(
+            'OLLAMA_MODEL', "llama3.2"
+            # 'OLLAMA_MODEL', "llama:7b"
+        )
+
         self.OLLAMA_EMBEDDINGS_MODEL = self.get_env(
-            'OLLAMA_EMBEDDINGS_MODEL', "llama:7b"
+            'OLLAMA_EMBEDDINGS_MODEL', "llama3.2"
+            # 'OLLAMA_EMBEDDINGS_MODEL', "llama:7b"
+        )
+
+        self.OLLAMA_TEMPERATURE = self.get_env(
+            'OLLAMA_TEMPERATURE', "0"
+        )
+
+        self.OLLAMA_BASE_URL = self.get_env(
+            'OLLAMA_BASE_URL', ""
         )
 
         # MongoDB credentials and other parameters