+++ CHANGELOG.md
@@ -17,6 +17,23 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Breaks
 
 
+## Unreleased
+## 0.1.11 (2024-10-14)
+---
+
+### New
+Implement preamble model to run OpenAI o1-mini/o1-preview models with Tools and System messages [GS-140].
+Add AI_PREAMBLE_MODEL_DEFAULT_TYPE, AI_PREAMBLE_MODEL_DEFAULT_MODEL, AI_PREAMBLE_MODEL_BASE_CONF, AI_PREAMBLE_MODEL_CUSTOM_CONF to customize the preamble model [GS-140].
+Implement ollama server [GS-139].
+Add AI_MODEL_ALLOW_SYSTEM_MSG, AI_MODEL_ALLOW_TOOLS, and AI_MODEL_NEED_PREAMBLE to manage models like Ollama "llava" that doesn't accept Tools [GS-140].
+
+### Changes
+Update ChatOllama adding the "langchain-ollama" dependency [GS-139].
+
+### Fixes
+Fix Tools make Agent returns empty responses in LCEL chains. Now Agent returns the result when there're no more Tools to call [GS-143].
+
+
 ## 0.1.10 (2024-10-07)
 ---